+++ includes/mappers/mapper_001.h
@@ -19,11 +19,11 @@ typedef struct Registers_001
    uint8_t PRG_bank;
 } Registers_001;
 
-cartridge_access_mode_t mapper001_cpu_read(nes_header_t *header, uint16_t position, uint16_t *mapped_addr, void* internal_registers);
-cartridge_access_mode_t mapper001_cpu_write(nes_header_t *header, uint16_t position, uint8_t data, uint16_t *mapped_addr, void* internal_registers);
+cartridge_access_mode_t mapper001_cpu_read(nes_header_t *header, uint16_t position, size_t *mapped_addr, void* internal_registers);
+cartridge_access_mode_t mapper001_cpu_write(nes_header_t *header, uint16_t position, uint8_t data, size_t *mapped_addr, void* internal_registers);
 
-cartridge_access_mode_t mapper001_ppu_read(nes_header_t *header, uint16_t position, uint16_t *mapped_addr, void* internal_registers);
-cartridge_access_mode_t mapper001_ppu_write(nes_header_t *header, uint16_t position, uint16_t *mapped_addr, void* internal_registers);
+cartridge_access_mode_t mapper001_ppu_read(nes_header_t *header, uint16_t position, size_t *mapped_addr, void* internal_registers);
+cartridge_access_mode_t mapper001_ppu_write(nes_header_t *header, uint16_t position, size_t *mapped_addr, void* internal_registers);
 
 void mapper001_init(nes_header_t* header, void* internal_registers);
